requirements.txt
openpyxl>=3.1.0
python-docx>=1.1.0
Pillow>=10.0.0